Remove `Frame` object and `stringify` decoder option.
Decoded frames are now simply objects. This simplifies the design and is
consistent with the idea that a single frame can't be extracted from,
therefore the `Extractor` will only work when piped to. This also
implies that it can infer the link type from the source stream without
us needing to copy it into each decoded frame.

The upside is performance and simplicity. The downside is less
flexibility, in particular we can't add methods to a frame object
anymore (e.g. pretty printing).

The `Decoder` now has a `decode` static method which takes in a buffer
and a link type as convenience to decode a single buffer (i.e. without
using the stream).