Change link type to be an integer.
Also add `capture.getLinkInfo` function to get back the name and a
description of this link type.

This change simplifies the interactions with the decoder (which requires
a integer, so the previous approach would have required us to implement
a name -> integer translation, which is difficult since link types are
defined as constants in pcap headers).